FC: initial check for fdt and its containing buffer
It is not inherently an error to have a containing buffer larger
than the actual FDT. FDTs can be grown and hence a larger buffer
may just accomodate expected growths.

However, a buffer that is actually smaller than the value in the
FDT header does not represent a valid situation and it is better
to bail off early.